Modifications to improve the structure
All the requests made by the previous commit's comments have been corrected.

The whitespaces have been removed, some names have been changed to conform to the context, the routes and collection are now the same as in "challenge 1", the "with" structures with a case have been replaced by case structures.