Code Increments
All the requests made by the previous commit's comments have been corrected.

The error blocks in the case structures are now written error -> error, the name of the product_exists? method is now fetch_product, the case structures in the update and delete methods now no longer have to search for the product before changing it, the project port is now 4000, some unused variables and methods have been removed, the halt() method has been added to the end of the pipeline of the fetch_product method, the update and delete methods now renspond with status 204.