Improve the structure and error handling
Based on some comments as a suggestion some modifications regarding the structure of the application and the addition of better error handling.

The modules LogstashController and RedisController were changed to ElasticSearchHelper and RedisHelper, the set, get and delete functions of RedisHelper file now have the _products suffix, the castore dependency was removed, the index used in the logs saving is now requests, the redis database now has error handling in case the database is not available and the update of the database data is done only in the endpoint index, now the log also stores the date and time when the request was done.