General updates
In the config file the name of the task_bunny queue has been changed to "reports".

In the ReportJob module the perform method parameter is now %{"type" => "products"}.

In the products file it is now possible to create a product with a predefined id.

In the data_case file now instead of using Repo.delete_all, Ecto.truncate will be used and the on_exit callback will no longer be used.