add mailer
A new application added, for sending e-mails.
The application has a MailerService module that is responsible for creating the structure of the email to be sent, including the attached report. The other module is MailerController, which will use the MailerService's email creation and send the email, every time a request is made in the endpoint /mailer.